# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RUMAAN ONLINE STORE API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  imageUrl: string;
  categoryId: number;
  categoryName: string;
  featured: boolean;
  inStock: boolean;
  whatsappNumber: string;
  createdAt: string;
}

export interface CreateProductInput {
  name: string;
  description: string;
  price: number;
  imageUrl: string;
  categoryId: number;
  featured?: boolean;
  inStock?: boolean;
  whatsappNumber: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  imageUrl: string;
  allowOnlinePurchase: boolean;
  showPrices: boolean;
  productCount: number;
}

export interface CreateCategoryInput {
  name: string;
  slug: string;
  description: string;
  imageUrl: string;
  allowOnlinePurchase?: boolean;
  showPrices?: boolean;
}

export interface Order {
  id: number;
  productId: number;
  productName: string;
  productPrice: number;
  receiptId: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  status: string;
  createdAt: string;
}

export interface CreateOrderInput {
  productId: number;
  receiptId: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
}

export interface UpdateOrderStatusInput {
  status: string;
}

export interface Review {
  id: number;
  productId: number;
  productName: string;
  reviewerName: string;
  rating: number;
  comment: string;
  imageUrls: string[];
  createdAt: string;
}

export interface CreateReviewInput {
  productId: number;
  reviewerName: string;
  rating: number;
  comment: string;
  imageUrls?: string[];
}

export interface AdminStats {
  totalProducts: number;
  totalOrders: number;
  totalReviews: number;
  pendingOrders: number;
  totalCategories: number;
}

export type GetProductsParams = {
  categoryId?: number;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  featured?: boolean;
};

export type GetReviewsParams = {
  productId?: number;
};