# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RUMAAN ONLINE STORE API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product management
  - name: categories
    description: Category management
  - name: orders
    description: Order management
  - name: reviews
    description: Product reviews
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get all orders (admin)
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Submit an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /reviews:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: Get reviews
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [products]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        featured:
          type: boolean
        inStock:
          type: boolean
        whatsappNumber:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - price
        - imageUrl
        - categoryId
        - categoryName
        - featured
        - inStock
        - whatsappNumber
        - createdAt

    CreateProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        categoryId:
          type: integer
        featured:
          type: boolean
        inStock:
          type: boolean
        whatsappNumber:
          type: string
      required:
        - name
        - description
        - price
        - imageUrl
        - categoryId
        - whatsappNumber

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        allowOnlinePurchase:
          type: boolean
        showPrices:
          type: boolean
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
        - description
        - imageUrl
        - allowOnlinePurchase
        - showPrices
        - productCount

    CreateCategoryInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        allowOnlinePurchase:
          type: boolean
        showPrices:
          type: boolean
      required:
        - name
        - slug
        - description
        - imageUrl

    Order:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productPrice:
          type: number
        receiptId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - productId
        - productName
        - productPrice
        - receiptId
        - customerName
        - customerPhone
        - customerAddress
        - status
        - createdAt

    CreateOrderInput:
      type: object
      properties:
        productId:
          type: integer
        receiptId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
      required:
        - productId
        - receiptId
        - customerName
        - customerPhone
        - customerAddress

    UpdateOrderStatusInput:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        reviewerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - productId
        - productName
        - reviewerName
        - rating
        - comment
        - imageUrls
        - createdAt

    CreateReviewInput:
      type: object
      properties:
        productId:
          type: integer
        reviewerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        imageUrls:
          type: array
          items:
            type: string
      required:
        - productId
        - reviewerName
        - rating
        - comment

    AdminStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        totalReviews:
          type: integer
        pendingOrders:
          type: integer
        totalCategories:
          type: integer
      required:
        - totalProducts
        - totalOrders
        - totalReviews
        - pendingOrders
        - totalCategories