# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RUMAAN ONLINE STORE API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalProducts: number;
  totalOrders: number;
  totalReviews: number;
  pendingOrders: number;
  totalCategories: number;
}