# lib/api-zod/src/generated/types/createOrderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RUMAAN ONLINE STORE API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderInput {
  productId: number;
  receiptId: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
}