# lib/api-zod/src/generated/types/createProductInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RUMAAN ONLINE STORE API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductInput {
  name: string;
  description: string;
  price: number;
  imageUrl: string;
  categoryId: number;
  featured?: boolean;
  inStock?: boolean;
  whatsappNumber: string;
}