# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RUMAAN ONLINE STORE API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  productId: number;
  productName: string;
  productPrice: number;
  receiptId: string;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  status: string;
  createdAt: string;
}